pyproject.toml
[tool.poetry]
name = "aiconverse"
version = "0.1.0"
description = "Console application to interact with an AI either through single-prompt inputs or an ongoing conversation using a REPL"
authors = ["Iwan van der Kleijn <iwanvanderkleijn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "0.3"  # Ensure Langchain is included
#ninja = "^1.10"    # For template handling
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
openai = "^1.45.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
load-dotenv = "^0.1.0"

[tool.poetry.scripts]
check-all = "aiconverse.run_checks:run_checks"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.11.2"
flake8 = "^7.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
